backend/requirements.txt
diffusers
transformers
accelerate
torch
torchvision
fastapi
uvicorn
python-multipart
pillow
scipy
safetensors
opencv-python-headless
numpy
